backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
reportlab==4.0.8
python-decouple==3.8
psycopg2-binary==2.9.9
